Created separate sending modules for groups and DMs.
